refactor: 테스트 픽스처 분리
